Feat (Player) Create Method and Route, first test OK
This route returns a Player (and his results). If the registered User had never played before, he will become a Player and will be added to the Players table, so he will be able to start playing and his results will be stored. If the registered User was already a Player, the Api will return also his data to the landing page, where he can start playing